+++ arm/ui/__init__.py
@@ -65,12 +65,14 @@ from arm.ui.auth.auth import route_auth  # noqa: E402,F811
 from arm.ui.database.database import route_database  # noqa: E402,F811
 from arm.ui.history.history import route_history  # noqa: E402,F811
 from arm.ui.jobs.jobs import route_jobs  # noqa: E402,F811
+from arm.ui.sendmovies.sendmovies import route_sendmovies  # noqa: E402,F811
 app.register_blueprint(route_settings)
 app.register_blueprint(route_logs)
 app.register_blueprint(route_auth)
 app.register_blueprint(route_database)
 app.register_blueprint(route_history)
 app.register_blueprint(route_jobs)
+app.register_blueprint(route_sendmovies)
 
 # Remove GET/page loads from logging
 import logging  # noqa: E402,F811
